# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/users": {
            "get": {
                "summary": "Obtener usuarios",
                "description": "Retorna una lista de todos los usuarios",
                "operationId": "1ad1cf4ffbfc84b02284335bdd828008",
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "nombre": {
                                                "type": "string",
                                                "example": "Juan Pérez"
                                            },
                                            "TipoUsuario": {
                                                "type": "string",
                                                "example": "cliente"
                                            },
                                            "correo": {
                                                "type": "string",
                                                "example": "juan.perez@example.com"
                                            },
                                            "telefono": {
                                                "type": "string",
                                                "example": "123456789"
                                            },
                                            "estado": {
                                                "type": "string",
                                                "example": "activo"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Crear un usuario",
                "description": "Crea un nuevo usuario en el sistema",
                "operationId": "5b6e72f178a035f354aff3d8b9eeb221",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Juan Pérez"
                                    },
                                    "TipoUsuario": {
                                        "type": "string",
                                        "example": "cliente"
                                    },
                                    "correo": {
                                        "type": "string",
                                        "example": "juan.perez@example.com"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "example": "123456789"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "banner": {
                                        "type": "string",
                                        "example": "https://example.com/banner.jpg"
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "example": "Calle 123"
                                    },
                                    "licencia": {
                                        "type": "string",
                                        "example": "ABC123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "example": "Usuario creado con éxito"
                                        },
                                        "request": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/{id}": {
            "put": {
                "summary": "Actualizar un usuario",
                "description": "Actualiza los datos de un usuario existente",
                "operationId": "619663f47b662de2164ec8e364e375ad",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Juan Pérez"
                                    },
                                    "TipoUsuario": {
                                        "type": "string",
                                        "example": "cliente"
                                    },
                                    "correo": {
                                        "type": "string",
                                        "example": "juan.perez@example.com"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "example": "123456789"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensaje": {
                                            "type": "string",
                                            "example": "Usuario actualizado"
                                        },
                                        "datos": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al actualizar el usuario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Hubo un error al actualizar el usuario"
                                        },
                                        "datos": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/{id}/estado": {
            "put": {
                "summary": "Actualizar el estado de un usuario",
                "description": "Actualiza el estado de un usuario existente",
                "operationId": "2e255783d76dbd9fd77a532b8b3a5b73",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "estado": {
                                        "type": "string",
                                        "example": "activo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Estado de usuario actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensaje": {
                                            "type": "string",
                                            "example": "Estado Actualizado con éxito"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al actualizar el estado del usuario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Hubo un error al actualizar el usuario"
                                        },
                                        "datos": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}